fix some missing renames
