+++ packages/insomnia-app/app/sync/__schemas__/type-schemas.ts
@@ -2,7 +2,7 @@ import { createBuilder, Schema } from '@develohpanda/fluent-builder';
 
 import { baseModelSchema } from '../../models/__schemas__/model-schemas';
 import { BackendProject, Branch, MergeConflict, SnapshotStateEntry, StatusCandidate, Team } from '../types';
-import { BackendProjectWithTeam } from '../vcs/normalize-project-team';
+import { BackendProjectWithTeam } from '../vcs/normalize-backend-project-team';
 
 export const projectSchema: Schema<BackendProject> = {
   id: () => 'id',
